renamed to DummyLogger and moved to eva.core.loggers
