+++ src/eva/core/loggers/__init__.py
@@ -1,0 +1,5 @@
+"""Loggers API."""
+
+from eva.core.loggers.dummy import DummyLogger
+
+__all__ = ["DummyLogger"]
