+++ src/eva/core/loggers/dummy.py
@@ -1,0 +1,32 @@
+import lightning.pytorch.loggers.logger
+
+
+class DummyLogger(lightning.pytorch.loggers.logger.DummyLogger):
+    """Dummy logger class.
+
+    This logger is currently used as a placeholder when saving results
+    to remote storage, as common lightning loggers do not work
+    with azure blob storage:
+    https://github.com/Lightning-AI/pytorch-lightning/issues/18861
+    https://github.com/Lightning-AI/pytorch-lightning/issues/19736
+
+    Simply disabling the loggers when pointing to remote storage doesn't work
+    because callbacks such as LearningRateMonitor or ModelCheckpoint require a
+    logger to be present.
+    """
+
+    def __init__(self, save_dir: str) -> None:
+        """Initializes the placeholder logger.
+
+        Args:
+            save_dir: The directory to save the logs.
+        """
+        super().__init__()
+        self._save_dir = save_dir
+
+    @property
+    def save_dir(self) -> str:
+        return self._save_dir
+
+    def __deepcopy__(self, memo=None):
+        return self
